Minor Changes
Default timeout is now 10 seconds for slow connections (the API call sometimes takes a few seconds
even on a decent connection anyway).

isConnected () now checks if the interface exists (this used to be all that was checked) and now if it has at
least one IPV4 address (is up).

ValueError exception is caught if the API response is malformed JSON (for some reason).